// src/risk/models/risk-rejection-code.enum.ts
export enum RiskRejectionCode {
  NO_ACTIVE_TRADING_PROFILE = 'NO_ACTIVE_TRADING_PROFILE',
  INVALID_TRADING_PROFILE = 'INVALID_TRADING_PROFILE',
  INVALID_ENTRY = 'INVALID_ENTRY',
  INVALID_STOP = 'INVALID_STOP',
  INVALID_RISK_PER_SHARE = 'INVALID_RISK_PER_SHARE',
  INVALID_PORTFOLIO_STATE = 'INVALID_PORTFOLIO_STATE',
  STOP_TOO_TIGHT = 'STOP_TOO_TIGHT',
  STOP_TOO_WIDE = 'STOP_TOO_WIDE',
  ENTRY_TOO_EXTENDED = 'ENTRY_TOO_EXTENDED',
  MAX_OPEN_TRADES_REACHED = 'MAX_OPEN_TRADES_REACHED',
  PORTFOLIO_RISK_LIMIT = 'PORTFOLIO_RISK_LIMIT',
  POSITION_SIZE_LIMIT = 'POSITION_SIZE_LIMIT',
  INSUFFICIENT_CAPITAL = 'INSUFFICIENT_CAPITAL',
  SECTOR_EXPOSURE_LIMIT = 'SECTOR_EXPOSURE_LIMIT',
  MINIMUM_RR_NOT_MET = 'MINIMUM_RR_NOT_MET',
  ZERO_RECOMMENDED_QUANTITY = 'ZERO_RECOMMENDED_QUANTITY',
  MISSING_REQUIRED_SETUP_CONTEXT = 'MISSING_REQUIRED_SETUP_CONTEXT',
}