// src/risk/models/risk-warning-code.enum.ts
export enum RiskWarningCode {
  SECTOR_METADATA_MISSING = 'SECTOR_METADATA_MISSING',
  OPEN_TRADE_SECTOR_UNAVAILABLE = 'OPEN_TRADE_SECTOR_UNAVAILABLE',
  CAPITAL_AVAILABILITY_APPROXIMATED = 'CAPITAL_AVAILABILITY_APPROXIMATED',
  TECHNICAL_TARGET_UNAVAILABLE = 'TECHNICAL_TARGET_UNAVAILABLE',
  PORTFOLIO_PRICE_DATA_STALE = 'PORTFOLIO_PRICE_DATA_STALE',
  PARTIAL_QUANTITY_UNAVAILABLE = 'PARTIAL_QUANTITY_UNAVAILABLE',
}